+++ R/scale-colour.R
@@ -93,6 +93,7 @@ scale_colour_continuous <- function(..., palette = NULL, aesthetics = "colour",
   palette <- if (!is.null(palette)) as_continuous_pal(palette)
   continuous_scale(
     aesthetics, palette = palette, guide = guide, na.value = na.value,
+    fallback.palette = pal_seq_gradient("#132B43", "#56B1F7"),
     ...
   )
 }
@@ -115,6 +116,7 @@ scale_fill_continuous <- function(..., palette = NULL, aesthetics = "fill", guid
   palette <- if (!is.null(palette)) as_continuous_pal(palette)
   continuous_scale(
     aesthetics, palette = palette, guide = guide, na.value = na.value,
+    fallback.palette = pal_seq_gradient("#132B43", "#56B1F7"),
     ...
   )
 }
@@ -137,6 +139,7 @@ scale_colour_binned <- function(..., palette = NULL, aesthetics = "colour", guid
   palette <- if (!is.null(palette)) pal_binned(as_discrete_pal(palette))
   binned_scale(
     aesthetics, palette = palette, guide = guide, na.value = na.value,
+    fallback.palette = pal_seq_gradient("#132B43", "#56B1F7"),
     ...
   )
 }
@@ -158,6 +161,7 @@ scale_fill_binned <- function(..., palette = NULL, aesthetics = "fill", guide = 
   palette <- if (!is.null(palette)) pal_binned(as_discrete_pal(palette))
   binned_scale(
     aesthetics, palette = palette, guide = guide, na.value = na.value,
+    fallback.palette = pal_seq_gradient("#132B43", "#56B1F7"),
     ...
   )
 }
@@ -227,6 +231,7 @@ scale_colour_discrete <- function(..., palette = NULL, aesthetics = "colour", na
   palette <- if (!is.null(palette)) as_discrete_pal(palette)
   discrete_scale(
     aesthetics, palette = palette, na.value = na.value,
+    fallback.palette = pal_hue(),
     ...
   )
 }
@@ -248,6 +253,7 @@ scale_fill_discrete <- function(..., palette = NULL, aesthetics = "fill", na.val
   palette <- if (!is.null(palette)) as_discrete_pal(palette)
   discrete_scale(
     aesthetics, palette = palette, na.value = na.value,
+    fallback.palette = pal_hue(),
     ...
   )
 }