+++ NEWS.md
@@ -9,8 +9,7 @@
   (@mitchelloharawild, #6609).
 * Fixed regression where `scale_{x,y}_*()` threw an error when an expression
   object is set to `labels` argument (@yutannihilation, #6617).
-
-
+* Improved palette fallback mechanism in scales (@teunbrand, #6669).
 * Allow `stat` in `geom_hline`, `geom_vline`, and `geom_abline`. (@sierrajohnson, #6559)
 * `stat_boxplot()` treats `width` as an optional aesthetic (@Yunuuuu, #6575)